requirements.txt
peewee
flask
pyyaml
